Add XML Document and move the Rune folder into the Text folder.
