+++ UnitTests/Drawing/StraightLineTests.cs
@@ -1,5 +1,4 @@
-﻿using System.Text;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Text;
 using Xunit;
